chore(worker): refactor entity-related modules to feature-related modules
