+++ worker/crates/action-universal/src/prelude.rs
@@ -7,10 +7,10 @@ pub use crate::attribute::sorter::AttributeSorter;
 pub use crate::color_converter::ColorConverter;
 pub use crate::dataframe::transformer::DataframeTransformer;
 pub use crate::dataframe::unions::DataframeUnion;
-pub use crate::entity::counter::EntityCounter;
-pub use crate::entity::filter::EntityFilter;
-pub use crate::entity::transformer::EntityTransformer;
 pub use crate::excel::writer::ExcelFileWriter;
+pub use crate::feature::counter::FeatureCounter;
+pub use crate::feature::filter::FeatureFilter;
+pub use crate::feature::transformer::FeatureTransformer;
 pub use crate::file::path_extractor::FilePathExtractor;
 pub use crate::file::reader::FileReader;
 pub use crate::file::writer::FileWriter;
